+++ hg/src/main/java/com/hg/web/rest/GoodsController.java
@@ -33,7 +33,7 @@ public class GoodsController {
     }
 
     @RequestMapping(value = "/goods/deleteGoods", method = RequestMethod.POST)
-    HgResponse deleteGoods(@RequestParam int id) {
+    HgResponse deleteGoods(@RequestParam long id) {
 
         goodsService.removeById(id);
         return HgResponse.success();
